+++ CHANGELOG.md
@@ -8,6 +8,11 @@
   config property `requestedSecretLifetime`. This helps reducing frequent Pod restarts ([#598]).
 - Run a `containerdebug` process in the background of each HBase container to collect debugging information ([#605]).
 - Aggregate emitted Kubernetes events on the CustomResources ([#612]).
+- Support configuring JVM arguments ([#620]).
+
+### Removed
+
+- BREAKING: The field `config.hbaseOpts` has been removed. Use JVM argument overrides instead to configure additional JVM arguments ([#620]).
 
 ### Changed
 
@@ -17,6 +22,7 @@
 [#605]: https://github.com/stackabletech/hbase-operator/pull/605
 [#611]: https://github.com/stackabletech/hbase-operator/pull/611
 [#612]: https://github.com/stackabletech/hbase-operator/pull/612
+[#620]: https://github.com/stackabletech/hbase-operator/pull/620
 
 ## [24.11.1] - 2025-01-09